+++ CHANGELOG.md
@@ -4,6 +4,10 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.3.2] - 2026-01-03
+### Added
+- Implemented automatic version syncing between `config.yaml` and the Python script at runtime.
+
 ## [1.3.1] - 2026-01-03
 ### Changed
 - Fixed security issue where MQTT password was visible in debug logs.